LayoutTests/fast/forms/fieldset/focus-in-fieldset-disabled.html

<!DOCTYPE html>
<body>
<script src="../../js/resources/js-test-pre.js"></script>
<fieldset disabled>
  <input id=input1>
</fieldset>
<script>
var input1 = document.getElementById('input1');
debug('A form control in a disabled fieldset is not focusable:');
shouldBe('input1.focus(); document.activeElement', 'document.body');

debug('It becomes focusable if the fieldset is enabled:');
document.getElementsByTagName('fieldset')[0].disabled = false;
shouldBe('input1.focus(); document.activeElement', 'input1');
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
