third_party/blink/web_tests/virtual/layout_ng_fieldset/fast/forms/fieldset/focus-in-fieldset-disabled-expected.txt

A form control in a disabled fieldset is not focusable:
PASS input1.focus(); document.activeElement is document.body
It becomes focusable if the fieldset is enabled:
PASS input1.focus(); document.activeElement is input1
PASS successfullyParsed is true

TEST COMPLETE
